// src/types/enums.ts
export enum AuditStatus {
  CREATED = "CREATED",
  CONFIRMED = "CONFIRMED",
  PREPERATION = "PREPERATION",
  FIELDWORK = "FIELDWORK",
  REPORTING = "REPORTING",
  SUBMITTED = "SUBMITTED",
  RESOLVED = "RESOLVED",
  CLOSED = "CLOSED",
  CANCELLED = "CANCELLED",
}

export enum ResourceType {
  AUDITOR = "AUDITOR",
  PRIMARY_CONTACT_OPERATOR = "PRIMARY_CONTACT_OPERATOR",
  PRIMARY_CONTACT_LEAD = "PRIMARY_CONTACT_LEAD",
  AUDIT_CONTACT_OPERATOR = "AUDIT_CONTACT_OPERATOR",
  AUDIT_CONTACT_NON_OPERATOR = "AUDIT_CONTACT_NON_OPERATOR",
  MANAGER = "MANAGER",
  OBSERVER = "OBSERVER",
}

export enum ScopeType {
  CAPITAL_EXPENDITURES = "CAPITAL_EXPENDITURES",
  OPERATING_EXPENSES = "OPERATING_EXPENSES",
  PRODUCTION_ALLOCATIONS = "PRODUCTION_ALLOCATIONS",
  EQUALIZATIONS = "EQUALIZATIONS",
  FEES = "FEES",
  FEE_INCOME = "FEE_INCOME",
  MARKETING = "MARKETING",
  MANAGEMENT = "MANAGEMENT",
  ADMINISTRATION = "ADMINISTRATION",
}

export enum Role {
  USER = "USER",
  ADMIN = "ADMIN",
  SUPER_USER = "SUPER_USER",
}

export enum StakeholderType {
  OPERATOR = "OPERATOR",
  AUDIT_LEAD = "AUDIT_LEAD",
  AUDIT_PARTICIPANT = "AUDIT_PARTICIPANT",
  NON_OP_OWNER = "NON_OP_OWNER",
  SERVICE_PROVIDER = "SERVICE_PROVIDER",
}